fix(server): §50.4 narrow focuser_position via GetInt64
SQLite stores INTEGER as 64-bit; reading focuser_position with GetInt32
would throw InvalidCastException for a position above Int32.MaxValue.
Read GetInt64 and narrow explicitly to the int DTO. Track the unindexed
focuser_position predicate under the existing stats-perf index TODO.

+++ OpenAstroAra.Server/Services/SqliteStatsService.cs
@@ -156,7 +156,9 @@ public sealed class SqliteStatsService : IStatsService {
         await using (var reader = await cmd.ExecuteReaderAsync(ct)) {
             while (await reader.ReadAsync(ct)) {
                 var temp = reader.GetDouble(0);
-                var pos = reader.GetInt32(1);
+                // SQLite stores INTEGER as 64-bit; narrow explicitly so a focuser
+                // position above Int32.MaxValue can't throw InvalidCastException.
+                var pos = (int)reader.GetInt64(1);
                 var ts = DateTimeOffset.Parse(reader.GetString(2), CultureInfo.InvariantCulture);
                 samples.Add(new FocusTempPointDto(TemperatureC: temp, FocuserPosition: pos, Timestamp: ts));
                 temps.Add(temp);
